# product-api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    Product:
        description: structure for an API product with json tags
        properties:
            description:
                description: description for the product
                maxLength: 1000
                type: string
                x-go-name: Description
            id:
                description: id for the product
                format: int64
                minimum: 1
                type: integer
                x-go-name: ID
            name:
                description: name for the product
                maxLength: 255
                type: string
                x-go-name: Name
            price:
                description: price for the product
                format: float
                minimum: 0
                type: number
                x-go-name: Price
            sku:
                description: sku (stock keeping unit) for the product
                pattern: '[a-z]+-[a-z]+-[a-z]+'
                type: string
                x-go-name: SKU
        required:
            - id
            - name
            - price
            - sku
        type: object
        x-go-package: my-go-project/product-api/data
info:
    description: '# Documentation for Product API'
    title: Product API
    version: 1.0.0
paths:
    /products:
        get:
            description: Returns a list of products
            operationId: listProducts
            responses:
                "200":
                    $ref: '#/responses/productsResponse'
            tags:
                - products
        post:
            description: Adds a new product
            operationId: addProduct
            responses:
                "201":
                    $ref: '#/responses/productResponse'
                "422":
                    $ref: '#/responses/validationErrorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
    /products/{id}:
        delete:
            description: Deletes a product
            operationId: deleteProduct
            parameters:
                - description: The id of the product to delete from the data store
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    $ref: '#/responses/noContent'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
        put:
            description: Updates an existing product
            operationId: updateProduct
            parameters:
                - description: The id of the product to delete from the data store
                  format: int64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    $ref: '#/responses/noContent'
                "422":
                    $ref: '#/responses/validationErrorResponse'
                "501":
                    $ref: '#/responses/errorResponse'
            tags:
                - products
produces:
    - application/json
responses:
    errorResponse:
        description: Generic error message as a string
        schema:
            properties:
                message:
                    type: string
                    x-go-name: Message
            type: object
    productResponse:
        description: Data structure representing a single product
        schema:
            $ref: '#/definitions/Product'
    productsResponse:
        description: A list of products
        schema:
            items:
                $ref: '#/definitions/Product'
            type: array
    validationErrorResponse:
        description: Validation errors as an array of strings
        schema:
            properties:
                message:
                    type: string
                    x-go-name: Message
            type: object
schemes:
    - http
swagger: "2.0"